fix: reset step count cronjob
