+++ back_end/server/api/__init__.py
@@ -1,3 +1,4 @@
 from .common import app, db
 from .main import *
 from .account import *
+from .pedometer import *
