+++ back_end/server/cronjob
@@ -1,0 +1,2 @@
+* * * * * /usr/bin/python3 /app/cron.py >> /var/log/cron.log 2>&1
+
