+++ back_end/server/requirements.txt
@@ -19,6 +19,7 @@ PyJWT==2.8.0
 pyodbc==4.0.39
 python-dotenv==1.0.0
 requests==2.31.0
+ruff==0.0.291
 SQLAlchemy==2.0.21
 typing_extensions==4.8.0
 urllib3==2.0.5
